PdfName/PdfString: Changed GetLength() to be unsigned
Other types already limits size types to unsigned which is 4 bytes also on 64bit

+++ src/pdfmm/base/PdfName.cpp
@@ -237,13 +237,13 @@ const string& PdfName::GetString() const
         return *m_utf8String;
 }
 
-size_t PdfName::GetLength() const
+unsigned PdfName::GetLength() const
 {
     expandUtf8String();
     if (m_utf8String == nullptr)
-        return m_data->length();
+        return (unsigned)m_data->length();
     else
-        return m_utf8String->length();
+        return (unsigned)m_utf8String->length();
 }
 
 const string& PdfName::GetRawData() const
